refactor(theme-classic): prepare to extract multiple items
